hcwiley_art/requirements.txt
Django==1.5
Pillow
MySQL-python==1.2.3
South==0.7.6
python-memcached
pyjade==1.5.2
django-pipeline
django-admin-bootstrapped
dj-database-url
django-tinymce
